pyproject.toml
[tool.poetry]
name = "verselect"
version = "0.0.1"
description = ""
authors = ["Bogdan Evstratenko <evstrat.bg@gmail.com>"]
repository = "https://github.com/team-monite/verselect"
classifiers = [
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python",
    "Topic :: Internet",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development",
    "Typing :: Typed",
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: AsyncIO",
    "Framework :: FastAPI",
    "Framework :: Pydantic",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Internet :: WWW/HTTP",
]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"
fastapi = "^0.103.1"
jinja2 = "^3.1.2"
uvicorn = {version = "^0.23.2", extras = ["standard"]}
python-multipart = "^0.0.6" # For FastAPI
httpx = "^0.25.0"
msgpack = ">=1.0.6"
monite-common = {version = ">=3.8.0", source = "monite"}
monite-logger = {version = ">=0.5.5", source = "monite"}
monite-requester = {version = ">=2.0.0", source = "monite"}
starlette-exporter = "^0.16.0"

[tool.poetry.group.dev.dependencies]
monite-ci = {version = ">=0.2.0", source = "monite"}
cadwyn = "^2.0.2"
black = "^23.11.0"
ruff = "^0.1.4"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"

[tool.coverage.run]
data_file = "coverage/coverage"
omit = [".venv/*"]
branch = true  # Enable branch coverage which is stricter than basic coverage

[tool.diff_cover]
fail_under = 100
compare_branch = "origin/master"

[tool.coverage.report]
fail_under = 100
skip_covered = true
skip_empty = true
omit = [".venv/*"]
exclude_lines = [
    "pragma: no cover",
    "assert_never\\(",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if False:",
    "if 0:",
    "class .*\\bProtocol\\):",
    "if __name__ == .__main__.:",
    # Impossible to cover
    # It is impossible to cover "if TYPE_CHECKING" statements because they never actually run
    "if TYPE_CHECKING:",
    "@(abc\\.)?abstractmethod",
    "@(typing\\.)?overload",
    "__rich_repr__",
    "__repr__",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
log_level = "INFO"

[tool.pyright]
fail_under = 0
reportMissingImports = true
reportUnusedImport = true
ignore = [".venv", "migrations"]

[tool.black]
line-length = 120
target-version = ["py310"]


[tool.ruff]
line-length = 120
select = [
    "F",        # pyflakes
    "E",        # pycodestyle errors
    "W",        # pycodestyle warnings
    "C90",      # mccabe
    "I",        # isort
    "N",        # pep8-naming
    "UP",       # pyupgrade
    "YTT",      # flake8-2020
    "S",        # flake8-bandit
    "BLE",      # flake8-blind-except
    "FBT003",   # flake8-boolean-trap
    "B",        # flake8-bugbear
    "A",        # flake8-builtins
    "COM",      # flake8-commas
    "C4",       # flake8-comprehensions
    "T10",      # flake8-debugger
    "ISC",      # flake8-implicit-str-concat
    "G010",     # Logging statement uses warn instead of warning
    "G201",     # Logging .exception(...) should be used instead of .error(..., exc_info=True)
    "G202",     # Logging statement has redundant exc_info
    "INP",      # flake8-no-pep420
    "PIE",      # flake8-pie
    "T20",      # flake8-print
    "PYI",      # flake8-pyi
    "PT",       # flake8-pytest-style
    "Q",        # flake8-quotes
    "RSE",      # flake8-raise
    "RET",      # flake8-return
    "SIM",      # flake8-simplify
    "TCH",      # flake8-type-checking
    "ARG",      # flake8-unused-arguments
    "PTH",      # flake8-use-pathlib
    "TD",       # flake8-todos
    "ERA",      # flake8-eradicate
    "PGH",      # pygrep-hooks
    "PLC0414",  # Import alias does not rename original package
    "PLE",      # Error
    "PLW",      # Warning
    "TRY",      # tryceratops
    "FLY",      # flynt
    "RUF",      # ruff-specific rules
    "ANN001",   # missing type annotation for arguments
    "ANN002",   # missing type annotation for *args
    "ANN003",   # missing type annotation for **kwargs
]
unfixable = [
    "ERA001"    # eradicate: found commented out code can be dangerous if fixed automatically
]
ignore = [
    "A002",   # builtin shadowing in arguments
    "A003",   # builtin shadowing in attributes
    "D203",   # 1 blank line required before class docstring
    "TD002",  # Missing author in TODO; try: `# TODO(<author_name>): ...`
    "ARG002", # Unused method argument
    "TRY003", # Avoid specifying long messages outside the exception class
    "TRY300", # Consider moving statement into the else clause
    "ARG001", # Unused first argument
    "ARG005", # Unused lambda argument
    "PT019",  # Fixture without value is injected as parameter, use @pytest.mark.usefixtures instead
    "SIM108"  # Use ternary operator instead of if-else block (ternaries lie to coverage)
]
extend-exclude = ["migrations"]

[tool.ruff.per-file-ignores]
"tests/*" = [
    "S",        # ignore bandit security issues in tests
    "B018",     # ignore assert in tests
    "PT012",    # ignore complex with pytest.raises clauses
]

[[tool.poetry.source]]
name = "monite"
url = "https://gitlab.monite.com/api/v4/projects/41/packages/pypi/simple"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
